Move sauce-platforms to the config directory.
