+++ automation-tests/config/sauce-platforms.js
