+++ automation-tests/scripts/run-all.js
@@ -14,7 +14,7 @@
  * should be tested.
  */
 
-const outputFormats = ["console", "json", "xunit"]; 
+const outputFormats = ["console", "json", "xunit"];
 
 
 var argv = require('optimist')
@@ -81,7 +81,7 @@ const path = require('path'),
       vows_path = path.join(__dirname, "../node_modules/.bin/vows"),
       vows_args = [(args.output === 'xunit') ? "--xunit" : "--json", "-i"],
       result_extension = process.env.RESULT_EXTENSION || "xml",
-      supported_platforms = require('../lib/sauce-platforms').platforms,
+      supported_platforms = require('../config/sauce-platforms').platforms,
       start_time = new Date().getTime(),
       glob = require('minimatch');
 
@@ -235,7 +235,7 @@ function startTesting() {
       }
 
       var stdOutReporter = null;
-      if (args.output === 'xunit') { 
+      if (args.output === 'xunit') {
         stdOutReporter = new FileReporter({
           output_path: outputPath
         });
